added LA and BEQ to default trees

+++ test_implementation/src/beq.rs
@@ -12,7 +12,7 @@ pub struct Bottleneck {
 }
 
 impl Bottleneck {
-    pub fn new(parent: &Vec<usize>, children: &Vec<Vec<usize>>, weights: &Vec<Vec<usize>>, root: usize) -> Self {
+    pub fn new(parent: &Vec<usize>, children: &Vec<Vec<usize>>, weights: &Vec<Vec<i64>>, root: usize) -> Self {
         let (c_tree, last_occ) = cartesian_on_tree(parent, children, weights, root);
         let (c_parent, c_children) = cartesian_to_tree(&c_tree);    // this returns parent and children of the cartesian tree while keeping the indices
         Self{
@@ -22,7 +22,7 @@ impl Bottleneck {
         }
     }
 
-    pub fn get(&self, a: usize, b: usize) -> usize {
+    pub fn get(&self, a: usize, b: usize) -> i64 {
         let ancestor = self.lca.get(self.last_occ[a], self.last_occ[b]);
         return self.c_tree[ancestor].weight;
     }
